[UNI-155] refactor : 기존 response DTO들의 패키지 이동
