+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/node/dto/response/SearchBuildingResDTO.java
@@ -1,4 +1,4 @@
-package com.softeer5.uniro_backend.node.dto;
+package com.softeer5.uniro_backend.node.dto.response;
 
 import java.util.List;
